requirements.txt
asgiref==3.9.1
attrs==25.3.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
diff-match-patch==20241021
Django==5.2.4
django-appconf==1.1.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-grappelli==4.0.2
django-import-export==4.3.8
djangorestframework==3.16.0
dnspython==2.7.0
drf-access-policy==1.5.0
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
pillow==11.3.0
pyasn1==0.6.1
pycparser==2.22
pycrypto==2.6.1
pymongo==4.13.2
pyparsing==3.2.3
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rest_condition==1.0.3
rpds-py==0.26.0
shortuuid==1.0.13
sqlparse==0.5.3
tablib==3.8.0
uritemplate==4.2.0
urllib3==2.5.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
whitenoise==6.9.0
celery>=5.5.0
redis>=5.0.0
qrcode[pil]==7.4.2
openpyxl>=3.1.2
flower==2.0.1